+++ src/main/java/raisetech/student/management/repository/StudentRepository.java
@@ -1,35 +1,34 @@
 package raisetech.student.management.repository;
 
 import java.util.List;
-import org.apache.ibatis.annotations.*;
+import org.apache.ibatis.annotations.Mapper;
+import org.apache.ibatis.annotations.Param;
 import raisetech.student.management.data.Student;
 
 /**
- * 受講生情報に関するデータベース操作を行うMyBatisリポジトリ。
+ * 受講生情報に関するデータベース操作を行うMyBatisリポジトリインターフェースです。
+ * <p>
+ * 受講生の登録、更新、検索、削除などを提供し、XMLマッパーファイルでSQLを制御します。
  */
 @Mapper
 public interface StudentRepository {
 
   /**
-   * 受講生IDで1件の受講生情報を取得します。
+   * 動的な検索条件（ふりがな、削除状態）に基づいて受講生情報を検索します。
    *
-   * @param studentId 検索する受講生ID
-   * @return 該当する受講生（存在しない場合は null）
+   * @param furigana        ふりがなによる部分一致検索条件（nullまたは空文字は無視）
+   * @param includeDeleted  論理削除された受講生も含めるかどうか
+   * @param deletedOnly     論理削除された受講生のみ取得するかどうか
+   * @return 条件に一致する受講生情報のリスト
    */
-  @Select("SELECT * FROM students WHERE student_id = #{studentId}")
-  @Results({
-      @Result(property = "studentId", column = "student_id"),
-      @Result(property = "fullName", column = "full_name"),
-      @Result(property = "furigana", column = "furigana"),
-      @Result(property = "nickname", column = "nickname"),
-      @Result(property = "email", column = "email"),
-      @Result(property = "location", column = "location"),
-      @Result(property = "age", column = "age"),
-      @Result(property = "gender", column = "gender"),
-      @Result(property = "remarks", column = "remarks"),
-      @Result(property = "deleted", column = "is_deleted"),
-      @Result(property = "deletedAt", column = "deleted_at")
-  })
+  List<Student> searchStudents(@Param("furigana") String furigana, @Param("includeDeleted") boolean includeDeleted, @Param("deletedOnly") boolean deletedOnly);
+
+  /**
+   * 受講生IDで受講生情報を取得します。
+   *
+   * @param studentId 受講生ID
+   * @return 該当する受講生情報（存在しない場合は null）
+   */
   Student findById(String studentId);
 
   /**
@@ -37,10 +36,6 @@ public interface StudentRepository {
    *
    * @param student 登録する受講生情報
    */
-  @Insert("""
-      INSERT INTO students (student_id, full_name, furigana, nickname, email, location, age, gender, remarks, created_at, is_deleted)
-      VALUES (#{studentId}, #{fullName}, #{furigana}, #{nickname}, #{email}, #{location}, #{age}, #{gender}, #{remarks}, now(), #{deleted})
-      """)
   void insertStudent(Student student);
 
   /**
@@ -48,76 +43,51 @@ public interface StudentRepository {
    *
    * @param student 更新する受講生情報
    */
-  @Update("""
-      UPDATE students SET full_name = #{fullName}, furigana = #{furigana}, nickname = #{nickname},
-      email = #{email}, location = #{location}, age = #{age}, gender = #{gender}, remarks = #{remarks},
-      is_deleted = #{deleted}, deleted_at = #{deletedAt} WHERE student_id = #{studentId}
-      """)
   void updateStudent(Student student);
 
   /**
-   * 削除されていない全ての受講生を取得します。
+   * 論理削除されていない受講生情報を取得します。
    *
-   * @return 論理削除されていない受講生リスト
+   * @return 有効な受講生情報のリスト
    */
-  @Select("SELECT * FROM students WHERE is_deleted = false")
   List<Student> searchActiveStudents();
 
   /**
-   * 論理削除フラグに関係なく、すべての受講生を取得します。
+   * 論理削除を含むすべての受講生情報を取得します。
    *
-   * @return 受講生エンティティのリスト
+   * @return 全受講生情報のリスト
    */
-  @Select("SELECT * FROM students")
-  @Results(id = "studentResultMap", value = {
-      @Result(property = "studentId", column = "student_id"),
-      @Result(property = "fullName", column = "full_name"),
-      @Result(property = "furigana", column = "furigana"),
-      @Result(property = "nickname", column = "nickname"),
-      @Result(property = "email", column = "email"),
-      @Result(property = "location", column = "location"),
-      @Result(property = "age", column = "age"),
-      @Result(property = "gender", column = "gender"),
-      @Result(property = "remarks", column = "remarks"),
-      @Result(property = "createdAt", column = "created_at"),
-      @Result(property = "deletedAt", column = "deleted_at"),
-      @Result(property = "deleted", column = "is_deleted")
-  })
   List<Student> findAllStudents();
 
   /**
-   * 論理削除された受講生のみを取得します。
+   * 論理削除された受講生情報のみを取得します。
    *
-   * @return 削除済み受講生のリスト
+   * @return 削除済み受講生情報のリスト
    */
-  @Select("SELECT * FROM students WHERE is_deleted = true")
   List<Student> findDeletedStudents();
 
   /**
-   * 指定されたふりがなに部分一致する受講生を検索します。
+   * ふりがなで部分一致検索し、論理削除されていない受講生を取得します。
    *
-   * @param furigana 検索対象のふりがな文字列
-   * @return 一致する受講生リスト
+   * @param furigana 検索するふりがな（部分一致）
+   * @return 条件に一致する受講生情報のリスト
    */
-  @Select("SELECT * FROM students WHERE furigana LIKE CONCAT('%', #{furigana}, '%') AND is_deleted = false")
-  List<Student> findByFurigana(String furigana);
+  List<Student> findByFurigana(@Param("furigana") String furigana);
 
   /**
-   * ふりがなで部分一致検索（論理削除された学生も含む）。
+   * ふりがなで部分一致検索し、論理削除状態を問わず受講生情報を取得します。
    *
-   * @param furigana 検索対象のふりがな
-   * @return 一致する学生のリスト
+   * @param furigana 検索するふりがな（部分一致）
+   * @return 条件に一致する受講生情報のリスト
    */
-  @Select("SELECT * FROM students WHERE furigana LIKE CONCAT('%', #{furigana}, '%')")
   List<Student> findByFuriganaIncludingDeleted(@Param("furigana") String furigana);
 
   /**
-   * 指定されたふりがなに一致する論理削除された受講生のみを検索します。
+   * ふりがなで部分一致検索し、論理削除された受講生情報のみを取得します。
    *
-   * @param furigana 検索対象のふりがな（部分一致）
-   * @return 条件に一致する削除済み受講生のリスト
+   * @param furigana 検索するふりがな（部分一致）
+   * @return 条件に一致する削除済み受講生情報のリスト
    */
-  @Select("SELECT * FROM students WHERE furigana LIKE CONCAT('%', #{furigana}, '%') AND is_deleted = true")
   List<Student> findDeletedStudentsByFurigana(@Param("furigana") String furigana);
 
   /**
@@ -125,9 +95,9 @@ public interface StudentRepository {
    *
    * @param studentId 削除対象の受講生ID
    */
-  @Delete("DELETE FROM students WHERE student_id = #{studentId}")
   void deleteById(String studentId);
 }
+
