+++ src/main/resources/application.properties
@@ -14,6 +14,7 @@ spring.jpa.hibernate.ddl-auto=update
 # MyBatis
 mybatis.configuration.map-underscore-to-camel-case=true
 mybatis.configuration.log-impl=org.apache.ibatis.logging.stdout.StdOutImpl
+mybatis.mapper-locations=classpath:mappers/*.xml
 
 # HikariCP
 spring.datasource.hikari.connectionTimeout=20000
